combine interface into one file for both 32 & 8266
