+++ platformio.ini
@@ -17,15 +17,16 @@ monitor_speed = 115200
 [env:esp32doit-devkit-v1]
 platform = espressif32
 board = esp32doit-devkit-v1
-board_build.filesystem = spiffs
 lib_deps = SHA-1 Hash
--D USE_SPIFFS
+build_flags = 
+    -D USE_SPIFFS
 
 [env:nodemcuv2]
 platform = espressif8266
 board = nodemcuv2
 board_build.filesystem = littlefs
--D USE_LITTLEFS
+build_flags =
+    -D USE_LITTLEFS
 monitor_filters = default, esp8266_exception_decoder
